+++ includes/class-woo-subzero.php
@@ -80,6 +80,7 @@ final class WSZ_Woo_Subzero
                     'test_deferred_start_minutes' => 1,
                     'enable_test_cycle_notifications' => 'no',
                     'enable_stripe_tokens' => 'no',
+                    'enable_mollie_tokens' => 'no',
                     'enable_role_transitions' => 'no',
                     'active_user_role' => 'customer',
                     'inactive_user_role' => '',
@@ -195,6 +196,7 @@ final class WSZ_Woo_Subzero
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-paynl-payment-token.php';
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-paynl-gateway.php';
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-stripe-gateway.php';
+        require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-mollie-gateway.php';
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/class-wsz-payment-handler.php';
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-test-card-gateway.php';
         require_once WSZ_WOO_SUBZERO_PATH . 'src/Payment/Gateway/class-wsz-tokenized-gateway.php';